refactor: 動画内時刻の型を見直して、int で統一するようにしました。
- positionDelta の型を int に修正しました。

+++ domain/learning/learning_session.py
@@ -66,7 +66,7 @@ class LearningSession:
         occurred_at: datetime,
         video_position: int,
         action: ViewingAction,
-        position_delta: float,
+        position_delta: int,
     ) -> LearningSession:
         event = ViewingEvent.create(
             id=event_id,
